+++ internal/cli/root.go
@@ -91,6 +91,7 @@ func NewRootCommand(info BuildInfo) *cobra.Command {
 	root.AddCommand(commands.NewSearchCommand())
 	root.AddCommand(commands.NewCatalogCommand())
 	root.AddCommand(commands.NewDependenciesCommand())
+	root.AddCommand(commands.NewInterfaceCommand())
 
 	return root
 }